move types out of UploadQueue
